models added, database bugfix

+++ database/migrations/2023_11_27_164615_create_trade_history_table.php
@@ -13,9 +13,9 @@ return new class extends Migration
     {
         Schema::create('trade_history', function (Blueprint $table) {
             $table->id();
-            $table->integer('painting_id');
-            $table->integer('buyer_id');
-            $table->integer('seller_id');
+            $table->unsignedBigInteger('painting_id');
+            $table->unsignedBigInteger('buyer_id');
+            $table->unsignedBigInteger('seller_id');
             $table->integer('trade_amount');
             $table->timestamps();
